src/event_seg/requirements.txt
accelerate==1.4.0
addict==2.4.0
aiofiles==22.1.0
aiosqlite==0.21.0
alembic==1.14.1
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
attrs==25.1.0
av==10.0.0
babel==2.17.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
blis==1.2.0
braceexpand==0.1.7
# Brotli @ file:///croot/brotli-split_1736182456865/work
cachetools==5.5.2
catalogue==2.0.10
# certifi @ file:///croot/certifi_1738623731865/work/certifi
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==2.1.1
click==8.1.8
cloudpathlib==0.20.0
colorlog==6.9.0
comm==0.2.2
confection==0.1.5
contexttimer==0.3.3
contourpy==1.3.0
cycler==0.12.1
cymem==2.0.11
debugpy==1.8.12
decorator==4.4.2
decord==0.6.0
defusedxml==0.7.1
deprecation==2.1.0
diffusers==0.16.0
distlib==0.3.9
easydict==1.9
einops==0.8.1
entrypoints==0.4
exceptiongroup==1.2.2
executing==2.2.0
fairscale==0.4.4
fastjsonschema==2.21.1
# filelock @ file:///croot/filelock_1700591183607/work
fonttools==4.56.0
fqdn==1.5.1
fsspec==2024.6.1
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.44
# gmpy2 @ file:///croot/gmpy2_1738085463648/work
greenlet==3.1.1
h5py==3.13.0
huggingface-hub==0.25.2
identify==2.6.8
idna==3.4
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
iopath==0.1.10
ipykernel==6.29.5
ipython==8.18.1
ipython-genutils==0.2.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-ydoc==0.2.5
jupyter_client==7.4.9
jupyter_core==5.7.2
jupyter_packaging==0.12.3
jupyter_server==2.15.0
jupyter_server_fileid==0.9.3
jupyter_server_terminals==0.5.3
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.8
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kaggle==1.6.17
kiwisolver==1.4.7
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.2
mkl-service==2.4.0
# mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
# mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
moviepy==1.0.3
# mpmath @ file:///croot/mpmath_1690848262763/work
murmurhash==1.0.12
narwhals==1.27.1
nbclassic==1.2.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
nltk==3.9.1
nodeenv==1.9.1
notebook==6.5.7
notebook_shim==0.2.4
numpy==1.26.4
omegaconf==2.3.0
open3d==0.14.1
opencv-python==4.7.0.72
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
optuna==4.2.1
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
peft==0.10.0
pexpect==4.9.0
Pillow==10.0.0
platformdirs==4.3.6
plotly==6.0.0
portalocker==3.1.1
pre_commit==4.1.0
preshed==3.0.9
proglog==0.1.10
prometheus_client==0.21.1
prompt_toolkit==3.0.50
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pycocoevalcap==1.2
pycocotools==2.0.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.1
# PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
python-magic==0.4.27
python-slugify==8.0.4
pytz==2025.1
# PyYAML @ file:///croot/pyyaml_1728657952215/work
pyyaml_env_tag==0.1
pyzmq==26.2.1
referencing==0.36.2
regex==2024.11.6
# requests @ file:///croot/requests_1730999120400/work
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.23.1
safetensors==0.5.2
scikit-image==0.22.0
scikit-learn==1.6.1
scipy==1.11.4
seaborn==0.13.2
Send2Trash==1.8.3
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.6
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.38
srsly==2.5.1
stack-data==0.6.3
streamlit==1.42.2
sympy==1.13.1
tenacity==9.0.0
terminado==0.18.1
text-unidecode==1.3
thinc==8.3.4
threadpoolctl==3.5.0
tifffile==2024.8.30
timm==0.4.12
tinycss2==1.4.0
tokenizers==0.13.3
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.2
torch==2.1.0+cu121
torchaudio==2.1.0+cu121
torchvision==0.16.0+cu121
tornado==6.4.2
tqdm==4.66.1
traitlets==5.14.3
transformers==4.33.2
triton==2.1.0
typer==0.15.1
types-python-dateutil==2.9.0.20241206
# typing_extensions @ file:///croot/typing_extensions_1734714854207/work
tzdata==2025.1
uri-template==1.3.0
urllib3==1.26.13
virtualenv==20.29.2
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webdataset==0.2.100
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.13
wrapt==1.17.2
y-py==0.6.2
ypy-websocket==0.8.4
zipp==3.21.0
